requirements.txt
flask==2.0.1
waitress==2.0.0
panel==0.12.1
python-dotenv==0.19.0
openai==0.27.0
